Fix a warning on 'len'.

+++ src/common.c
@@ -52,7 +52,7 @@ char *set_program_name (char *argv0)
   char   my_name [_MAX_PATH];
   DWORD  len = GetModuleFileName (NULL, my_name, sizeof(my_name));
 
-  snprintf (ret, sizeof(ret), "%.*s %s", len, my_name, argv0);
+  snprintf (ret, sizeof(ret), "%.*s %s", (int)len, my_name, argv0);
   program_name = ret;
   return (ret);
 }
